refactor: move loaders to separated file
